完成任务自动推进

+++ ZezeJava/ZezeJava/src/main/java/Zeze/Builtin/Game/TaskBase/BSubPhaseReadOnly.java
@@ -9,6 +9,6 @@ public interface BSubPhaseReadOnly {
 
     public long getSubPhaseId();
     public int getCompleteType();
-    public int getNextSubPhaseId();
+    public long getNextSubPhaseId();
     public Zeze.Transaction.Collections.PList2ReadOnly<Zeze.Builtin.Game.TaskBase.BTaskCondition, Zeze.Builtin.Game.TaskBase.BTaskConditionReadOnly> getConditionsReadOnly();
 }
